[PIL-1514-allow-test-users] Add a featureflag to allow test users in all environments except production (#77)
We have found that test users created via API Platform do not have a
credential role. We rely on the credential role to ensure a user is a
User (instead of Assistant), therefore this has posed a problem.

The solution we have chosen is to add a featureflag to allow test users,
which can be turned off in production