[PIL-1781-correct-order] Maintain order of attributes (#97)
Maintain order of attributes. This was an issue because anything created
in swagger.yaml doesn't come out in the same order.

This happened because the jsonMapper is parsing the JSON as:
```
val jsonMap    = jsonMapper.readValue(jsonString, classOf[Map[String, Any]])
```

This created all objects as `Map[String, Any]` which is not ordered.
This creates a random ordering of attributes that is carried forward
into the yaml.

**The fix**
Now, instead of using Jackson to parse into a Map and modify keys, we
are using Play Json transformation. This maintains the order of
attributes because the JSON isn't being parsed.