PIL-1781 - Add real examples of endpoint requests to OAS (#94)
This ticket was to enhance the OAS to be more helpful by providing
realistic examples of requests that external parties could send

### Changes
- added examples component to application.yaml
- referencing to use examples in the spec
- swagger.yml adjusted according to application.yaml updates
- testOnly application.yaml file updated

### Update
Further updates were to update the order of example request attributes.
Before Richard refactored the sbt task, the task would reorder the
attributes in the example request payload. This has since been fixed so
I had to correct the order the request for the last time in the
swagger.yml & run the task to generate a new version of the
1.0/application.yaml and testOnly/application.yaml


![image](https://github.com/user-attachments/assets/5f0b1d74-1212-4ce9-bda8-c57390474b57)

----->


![image](https://github.com/user-attachments/assets/26259188-6e06-4d96-95e9-28cd48fd42e6)

---------

Co-authored-by: Richard Dowsett <180321016+ridow-hmrc@users.noreply.github.com>
Co-authored-by: James Miller <7511223+JamesMMiller@users.noreply.github.com>